moved node related tests from test/unit to test/unit/node and updated paths to files to reflect the directory change
