+++ test/unit/node/Binary.spec.js
