+++ test/unit/node/coverage.spec.js
