+++ test/unit/node/reporters/BaseReporter.spec.js
